Add files via upload
The code is essentially a simple battle simulator where two actors take turns attacking each other until one wins. The battle details and the winner are displayed in an HTML element with the ID `"battle"`.

1. `RPGActor` Constructor Function
The `RPGActor` function defines an actor with three properties:
- life: The health points of the actor, randomly set between 10 and 20.
- hitpoints: The damage the actor can inflict, randomly set between 5 and 10.
- alive: A boolean that tracks whether the actor is still alive (initially set to `true`).

The `RPGActor` function also has a method called `attack`, which allows an actor to attack an enemy. The attack has a chance of success (two out of three), and if successful, it reduces the enemy’s life points by the attacking actor's hitpoints. If the enemy's life drops to 0 or below, the enemy is considered dead (`alive = false`).

 2. `play` Function
The `play` function sets up the battle between two actors (`actor1` and `actor2`) and displays the results on a webpage:

The output  shows a sequence of attacks until one actor's life reaches 0, after which a victory message is displayed.